supporting not round hours

+++ app/src/main/java/com/example/craftket/Models/TimeSlot.kt
@@ -2,8 +2,8 @@ package com.example.craftket.Models
 
 data class TimeSlot(
     val date: String = "",
-    val startTime: Int = 0,
-    val endTime: Int = 24,
+    val startTime: Float = 0f,
+    val endTime: Float = 24f,
     val capacity: Int = 0,
     val registeredUsers: List<String> = emptyList()
 )
